% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R94KarBargh1}
\alias{R94KarBargh1}
\title{Table R94KarBargh1 \cr
Worksheet related to miscellaneous incomes of rural household members in 1394 \cr
كاربرگ مربوط به درامدهاي متفرقه اعضاء خانوار روستایی سال 1394}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف فرد}
\item Description: Person row number
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{شماره ردیف فعالیت}
\item Description: Activity row number
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{درآمد}
\item Description: Income
}
}
}
\usage{
R94KarBargh1
}
\description{
{R for Rural} \cr
{94 for 1394}
}
\keyword{datasets}
