% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U94P3S11}
\alias{U94P3S11}
\title{Table U94P3S11 \cr
Household ready meals in hotels and restaurants expenditure of urban household in 1394 \cr
هزينه هاي  غذاهاي آماده هتل و رستوران خانوار شهری سال 1394}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
U94P3S11
}
\description{
{U for Urban} \cr
{94 for 1394} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
