% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U94P4S04}
\alias{U94P4S04}
\title{Table U94P4S04 \cr}
\format{
A data frame with 5 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description is provided for this table
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description is provided for this table
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description is provided for this table
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description is provided for this table
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description is provided for this table
}
}
}
\usage{
U94P4S04
}
\description{
{U for Urban} \cr
{94 for 1394} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
